modular design of bottomModalWithChoice
