+++ lib/screens/boards/posts/creation/post_creation_board.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:guam_community_client/commons/bottom_modal/bottom_modal_with_choice.dart';
-import 'package:guam_community_client/commons/custom_divider.dart';
 import 'package:guam_community_client/styles/colors.dart';
 import 'package:guam_community_client/styles/fonts.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
@@ -80,14 +79,58 @@ class _PostCreationBoardState extends State<PostCreationBoard> {
                 BottomModalWithChoice(
                   title: '게시판을 선택해주세요.',
                   back: '완료',
-                  choiceFunc: setBoardType,
-                  selectedChoice: widget.input['boardType'],
+                  children: [
+                    _boardType('익명게시판'),
+                    _boardType('자유게시판'),
+                    _boardType('구인게시판'),
+                    _boardType('정보공유게시판'),
+                    _boardType('홍보게시판')
+                  ],
                 ),
               ],
             ),
           )
         )
       )
+    );
+  }
+
+  Widget _boardType(String boardType) {
+    return Builder(
+      builder: (context) => InkWell(
+        onTap: () {
+          setBoardType(boardType);
+          Navigator.pop(context);
+        },
+        child: Padding(
+          padding: EdgeInsets.symmetric(vertical: 10),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Container(
+                height: 20,
+                child: Text(
+                  boardType,
+                  style: TextStyle(
+                    fontSize: 16,
+                    fontFamily: GuamFontFamily.SpoqaHanSansNeoRegular,
+                    color: boardType == widget.input['boardType']
+                        ? GuamColorFamily.purpleCore
+                        : GuamColorFamily.grayscaleGray3,
+                  ),
+                ),
+              ),
+              if (boardType == widget.input['boardType'])
+                IconButton(
+                  padding: EdgeInsets.only(right: 8),
+                  constraints: BoxConstraints(),
+                  icon: SvgPicture.asset('assets/icons/check.svg'),
+                  onPressed: null,
+                ),
+            ],
+          ),
+        ),
+      ),
     );
   }
 }
